+++ src/templates/index.ts
@@ -5,6 +5,9 @@ import { GetTemplatesByCollectionId } from './domain/useCases/GetTemplatesByColl
 import { GetTemplate } from './domain/useCases/GetTemplate'
 import { SetTemplateAsDefault } from './domain/useCases/SetTemplateAsDefault'
 import { UnsetTemplateAsDefault } from './domain/useCases/UnsetTemplateAsDefault'
+import { UpdateTemplateMetadata } from './domain/useCases/UpdateTemplateMetadata'
+import { UpdateTemplateLicenseTerms } from './domain/useCases/UpdateTemplateLicenseTerms'
+import { UpdateTemplateTermsOfAccess } from './domain/useCases/UpdateTemplateTermsOfAccess'
 
 const templatesRepository = new TemplatesRepository()
 
@@ -12,6 +15,9 @@ const createTemplate = new CreateTemplate(templatesRepository)
 const deleteTemplate = new DeleteTemplate(templatesRepository)
 const getTemplatesByCollectionId = new GetTemplatesByCollectionId(templatesRepository)
 const getTemplate = new GetTemplate(templatesRepository)
+const updateTemplateMetadata = new UpdateTemplateMetadata(templatesRepository)
+const updateTemplateLicenseTerms = new UpdateTemplateLicenseTerms(templatesRepository)
+const updateTemplateTermsOfAccess = new UpdateTemplateTermsOfAccess(templatesRepository)
 const setTemplateAsDefault = new SetTemplateAsDefault(templatesRepository)
 const unsetTemplateAsDefault = new UnsetTemplateAsDefault(templatesRepository)
 
@@ -21,8 +27,12 @@ export {
   getTemplatesByCollectionId,
   getTemplate,
   setTemplateAsDefault,
-  unsetTemplateAsDefault
+  unsetTemplateAsDefault,
+  updateTemplateMetadata,
+  updateTemplateLicenseTerms,
+  updateTemplateTermsOfAccess
 }
+
 export {
   CreateTemplateDTO,
   TemplateFieldDTO,
@@ -32,4 +42,6 @@ export {
   TemplateFieldValueControlledVocabularyDTO,
   TemplateInstructionDTO
 } from './domain/dtos/CreateTemplateDTO'
+export { UpdateTemplateMetadataDTO } from './domain/dtos/UpdateTemplateMetadataDTO'
+export { UpdateTemplateLicenseTermsDTO } from './domain/dtos/UpdateTemplateLicenseTermsDTO'
 export { Template, TemplateInstruction } from './domain/models/Template'